Integer parse code updated after some suggestions.

+++ addons/binding/org.openhab.binding.homematic/src/main/java/org/openhab/binding/homematic/internal/communicator/parser/CommonRpcParser.java
@@ -40,7 +40,7 @@ public abstract class CommonRpcParser<M, R> implements RpcParser<M, R> {
             return (Integer) object;
         }
         try {
-            return Long.valueOf(Double.valueOf(ObjectUtils.toString(object)).longValue()).intValue();
+            return Double.valueOf(ObjectUtils.toString(object)).intValue();
         } catch (NumberFormatException ex) {
             return null;
         }
@@ -181,7 +181,7 @@ public abstract class CommonRpcParser<M, R> implements RpcParser<M, R> {
         } else if (value.equalsIgnoreCase("false") || value.equalsIgnoreCase("off")) {
             return (Boolean.FALSE);
         } else if (value.matches("(-|\\+)?[0-9]+")) {
-            return Integer.valueOf((Long.valueOf(value).intValue()));
+            return Long.valueOf(value).intValue();
         } else if (value.matches("[-+]?[0-9]*\\.?[0-9]+([eE][-+]?[0-9]+)?")) {
             return (Double.valueOf(value));
         } else {
